backend/requirements.txt
# MotorSoft Backend Dependencies
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis
redis>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0

# File handling
aiofiles>=23.2.1
boto3>=1.34.0  # S3 compatible storage

# Excel/CSV parsing (WinOLS import)
pandas>=2.2.0
openpyxl>=3.1.2

# Telegram Bot
aiogram>=3.4.1

# Utils
httpx>=0.26.0
python-dotenv>=1.0.0
loguru>=0.7.2
